chore: 타입 파일 폴더링
