+++ frontend/src/hooks/levellog/types/index.ts
@@ -1,0 +1,5 @@
+export interface LevellogCustomHookType {
+  inputValue: string;
+  teamId: string | undefined;
+  levellogId: string | undefined;
+}
