+++ frontend/src/hooks/levellog/useLevellogEdit.ts
@@ -4,13 +4,14 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { useMutation } from '@tanstack/react-query';
 import { Editor } from '@toast-ui/react-editor';
 
+import { LevellogCustomHookType } from 'hooks/levellog/types';
+
 import useLevellogQuery from 'hooks/levellog/useLevellogQuery';
 import useSnackbar from 'hooks/utils/useSnackbar';
 
 import { MESSAGE } from 'constants/constants';
 
 import { requestEditLevellog } from 'apis/levellog';
-import { LevellogCustomHookType } from 'types/levellog';
 import { teamGetUriBuilder } from 'utils/util';
 
 const useLevellogEdit = () => {
