+++ frontend/src/hooks/preQuestion/types/index.ts
@@ -1,0 +1,9 @@
+import { PreQuestionFormatType } from 'types/preQuestion';
+
+export interface PreQuestionCustomHookType {
+  teamId: string;
+  levellogId: string | undefined;
+  preQuestionId: string | undefined;
+  preQuestion: PreQuestionFormatType;
+  preQuestionContent: string;
+}
