+++ frontend/src/hooks/team/types/index.ts
@@ -1,0 +1,12 @@
+export interface MembersCustomHookType {
+  nicknameValue: string;
+}
+
+export interface TeamCustomHookType {
+  title: string;
+  place: string;
+  startAt: string;
+  interviewerNumber: number;
+  watcherIds: string[];
+  participantIds: string[];
+}
