+++ frontend/src/types/interviewQuestion.ts
@@ -1,10 +1,3 @@
-export interface InterviewQuestionApiType {
-  accessToken: string | null;
-  levellogId: string | undefined;
-  interviewQuestionId: string;
-  interviewQuestion: string;
-}
-
 export interface InterviewQuestionInfoType {
   id: string;
   content: string;
@@ -29,9 +22,4 @@ export interface InterviewQuestionSearchResultType {
   like: boolean;
   likeCount: number;
 }
-export interface InterviewQuestionSearchApiType {
-  results: InterviewQuestionSearchResultType[];
-  page: number;
-}
-
 export type InterviewQuestionSort = 'likes' | 'latest' | 'oldest';
