--- frontend/src/types/login.ts
@@ -1,4 +1,0 @@
-export interface LoginApiType {
-  code: string;
-  accessToken: string;
-}
