+++ frontend/src/types/member.ts
@@ -1,12 +1,3 @@
-export interface MembersApiType {
-  accessToken: string | null;
-  nickname: string;
-}
-
-export interface MembersCustomHookType {
-  nicknameValue: string;
-}
-
 export interface MemberType {
   id: string;
   nickname: string;
