+++ frontend/src/types/preQuestion.ts
@@ -1,21 +1,5 @@
 import { ParticipantType } from 'types/team';
 
-export interface PreQuestionCustomHookType {
-  teamId: string;
-  levellogId: string | undefined;
-  preQuestionId: string | undefined;
-  preQuestion: PreQuestionFormatType;
-  preQuestionContent: string;
-}
-
-export interface PreQuestionApiType {
-  accessToken: string | null;
-  levellogId: string | undefined;
-  preQuestionId: string | undefined;
-  preQuestion: PreQuestionFormatType;
-  preQuestionContent: string;
-}
-
 export interface PreQuestionResultType {
   content: string;
 }
