--- frontend/src/types/uri.ts
@@ -1,7 +1,0 @@
-export interface UriCustomHookType {
-  teamId: string | undefined;
-  levellogId: string | undefined;
-  feedbackId: string | undefined;
-  preQuestionId: string | undefined;
-  authorId: string | undefined;
-}
